+++ src/mirage_flow.ml
@@ -27,64 +27,16 @@ let pp_or_eof d ppf = function
   | `Data a -> d ppf a
   | `Eof    -> Fmt.string ppf "End-of-file"
 
-module type ABSTRACT = sig
-  type +'a io
+module type S = sig
   type error
   val pp_error: error Fmt.t
-  type write_error
+  type nonrec write_error = private [> write_error ]
   val pp_write_error: write_error Fmt.t
-  type buffer
   type flow
-  val read: flow -> (buffer or_eof, error) result io
-  val write: flow -> buffer -> (unit, write_error) result io
-  val writev: flow -> buffer list -> (unit, write_error) result io
-  val close: flow -> unit io
-end
-
-module type S = ABSTRACT with type write_error = private [> write_error]
-  [@@ocaml.warning "-34"]
-
-module type CONCRETE = ABSTRACT
-  with type error = [`Msg of string]
-   and type write_error = [write_error | `Msg of string]
-
-module Concrete (S: S) (IO: sig
-       type 'a t = 'a S.io
-       val map: ('a -> 'b) -> 'a t -> 'b t
-  end) =
-struct
-  type 'a io = 'a S.io
-  type error = [`Msg of string]
-  type write_error = [ `Closed | `Msg of string]
-  type buffer = S.buffer
-  type flow = S.flow
-
-  let pp_error ppf = function
-    | `Msg s -> Fmt.string ppf s
-
-  let pp_write_error ppf = function
-    | #error as e -> pp_error ppf e
-    | `Closed     -> pp_write_error ppf `Closed
-
-  let lift_read = function
-    | Ok x    -> Ok x
-    | Error e -> Error (`Msg (Fmt.strf "%a" S.pp_error e))
-
-  let lift_write = function
-    | Ok ()         -> Ok ()
-    | Error `Closed -> Error `Closed
-    | Error e       -> Error (`Msg (Fmt.strf "%a" S.pp_write_error e))
-
-  let read t = IO.map lift_read (S.read t)
-  let write t b = IO.map lift_write (S.write t b)
-  let writev t bs = IO.map lift_write (S.writev t bs)
-  let close t = S.close t
-end
-
-module type SHUTDOWNABLE = sig
-  include S
-  val shutdown_write: flow -> unit io
-  val shutdown_read: flow -> unit io
+  val read: flow -> (Cstruct.t or_eof, error) result Lwt.t
+  val write: flow -> Cstruct.t -> (unit, write_error) result Lwt.t
+  val writev: flow -> Cstruct.t list -> (unit, write_error) result Lwt.t
+  val close: flow -> unit Lwt.t
 end
 
 type stats = {